# feat: Optimise `load_policy_line` to avoid quadratic individual-character loop