--- a/casbin/persist/adapter.py
+++ b/casbin/persist/adapter.py
@@ -12,6 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import re
+
+_INTERESTING_TOKENS_RE = re.compile(r"[,\[\]\(\)]")
+
 
 def _extract_tokens(line):
     """Return the list of 'tokens' from the line, or None if this line has none"""
@@ -24,20 +28,29 @@ def _extract_tokens(line):
 
     stack = []
     tokens = []
-    for c in line:
+
+    # The tokens are separated by commas, but we support nesting so a naive `line.split(",")` is
+    # wrong. E.g. `abc(def, ghi), jkl` is two tokens: `abc(def, ghi)` and `jkl`. We do this by
+    # iterating over the locations of any tokens of interest, and either:
+    #
+    # - [](): adjust the nesting depth
+    # - ,: slice the line to save the token, if the , is at the top-level, outside all []()
+    #
+    # `start_idx` represents the start of the current token, that we haven't seen a `,` for yet.
+    start_idx = 0
+    for match in _INTERESTING_TOKENS_RE.finditer(line):
+        c = match.group()
         if c == "[" or c == "(":
             stack.append(c)
-            tokens[-1] += c
         elif c == "]" or c == ")":
             stack.pop()
-            tokens[-1] += c
         elif c == "," and len(stack) == 0:
-            tokens.append("")
-        else:
-            if len(tokens) == 0:
-                tokens.append(c)
-            else:
-                tokens[-1] += c
+            # we've found the end of a top level token so save that and start a new one
+            tokens.append(line[start_idx : match.start()])
+            start_idx = match.end()
+
+    # trailing token after the last ,
+    tokens.append(line[start_idx:])
 
     tokens = [x.strip() for x in tokens]
     return tokens