--- a/casbin/persist/adapter.py
+++ b/casbin/persist/adapter.py
@@ -44,8 +44,9 @@ def _extract_tokens(line):
             stack.append(c)
         elif c == "]" or c == ")":
             stack.pop()
-        elif c == "," and len(stack) == 0:
-            # we've found the end of a top level token so save that and start a new one
+        elif len(stack) == 0:
+            # must be a comma outside of any nesting: we've found the end of a top level token so
+            # save that and start a new one
             tokens.append(line[start_idx : match.start()])
             start_idx = match.end()