This dramatically improves the performance of the `casbin.persist.adapter.load_policy_line` function:

- before: iterate over characters individually and build each token string by `+=`'ing those characters. The latter in particular leads to quadratic performance, based the length of the string (e.g. https://stackoverflow.com/a/34008199/1256624), which is noticeable with "long" lines, like if they contain UUIDs.
- after: use a regex to find the start/end index of each tokens and slice out each token string in one go.

This function was appearing very high in our profiles of initialising our adapters. We have two UUIDs appear in most of our lines.

To make this a safe refactoring, I've pulled out an internal function that does the string manipulation in isolation and tested it.

I've also added a benchmark that demonstrates the performance improvements. Here's the total change of the median time that `pytest-benchmark` reports before and after this change, in microseconds:

| Benchmark                                  | Before | After | Change |
|--------------------------------------------|-------:|------:|-------:|
| `test_benchmark_extract_tokens_long_nested`  |   27.0 |  2.46 |   -91% |
| `test_benchmark_extract_tokens_long_simple`  |   13.6 |  1.42 |   -90% |
| `test_benchmark_extract_tokens_short_nested` |   3.25 |  1.63 |   -50% |
| `test_benchmark_extract_tokens_short_simple` |   1.58 |  0.917 |   -42% |

That is, the "long" tests get 10× faster. Even longer lines will likely see a greater speed-up too.

---

Thanks for casbin!